WT-18068 Mark the local drop on the session
The pickup discard needs a drop that publishes nothing and releases no shared
storage, both of which belong to the node that removed the table. Carrying that
as an argument put a flag on every drop function and every caller, none of which
care about it.

Use a session flag instead. The layered drop reads it where the two decisions
are made, the recursion through the constituents carries it for free, and the
internal session the schema drop may switch to already inherits the session's
flags. The session flag field was full, so it widens to 64 bits; the locals that
name themselves after session flags but hold thread group flags keep their
type.

+++ src/log/log_mgr.c
@@ -1106,7 +1106,7 @@ __wt_logmgr_open(WT_SESSION_IMPL *session)
     WT_CONNECTION_IMPL *conn;
     WT_LOG_MANAGER *log_mgr;
     uint64_t now;
-    uint32_t session_flags;
+    uint64_t session_flags;
 
     conn = S2C(session);
     log_mgr = &conn->log_mgr;
