+++ src/session/session_api.c
@@ -2822,7 +2822,7 @@ __wt_open_session(WT_CONNECTION_IMPL *conn, WT_EVENT_HANDLER *event_handler, con
  */
 int
 __wt_open_internal_session(WT_CONNECTION_IMPL *conn, const char *name, bool open_metadata,
-  uint32_t session_flags, uint32_t session_lock_flags, WT_SESSION_IMPL **sessionp)
+  uint64_t session_flags, uint32_t session_lock_flags, WT_SESSION_IMPL **sessionp)
 {
     WT_SESSION_IMPL *session;
